my_agent/requirements.txt
langgraph
langchain_anthropic
tavily-python
langchain_community
langchain_openai
nltk
langsmith==0.1.141
traceloop-sdk
opentelemetry-instrumentation-redis